Moved tests into their appropriate files
