+++ app/backend/src/tests/test_api.py
@@ -4,7 +4,8 @@ import grpc
 import pytest
 from google.protobuf import empty_pb2, wrappers_pb2
 
-from couchers.db import session_scope
+from couchers import errors
+from couchers.db import get_user_by_field, session_scope
 from couchers.models import Complaint
 from couchers.utils import now, to_aware_datetime
 from pb import api_pb2, jail_pb2
@@ -139,6 +140,38 @@ def test_get_user(db):
         assert res.user_id == user2.id
         assert res.username == user2.username
         assert res.name == user2.name
+
+
+def test_get_invisible_user_by_username(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user(is_deleted=True)
+
+    with api_session(token1) as api:
+        with pytest.raises(grpc.RpcError) as e:
+            api.GetUser(api_pb2.GetUserReq(user=user2.username))
+    assert e.value.code() == grpc.StatusCode.NOT_FOUND
+    assert e.value.details() == errors.USER_NOT_FOUND
+
+
+def test_get_invisible_user_by_id(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user(is_deleted=True)
+    with api_session(token1) as api:
+        with pytest.raises(grpc.RpcError) as e:
+            api.GetUser(api_pb2.GetUserReq(user=str(user2.id)))
+    assert e.value.code() == grpc.StatusCode.NOT_FOUND
+    assert e.value.details() == errors.USER_NOT_FOUND
+
+
+def test_get_invisible_user_by_email(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user(is_deleted=True)
+
+    with api_session(token1) as api:
+        with pytest.raises(grpc.RpcError) as e:
+            api.GetUser(api_pb2.GetUserReq(user=user2.email))
+    assert e.value.code() == grpc.StatusCode.NOT_FOUND
+    assert e.value.details() == errors.USER_NOT_FOUND
 
 
 def test_update_profile(db):
@@ -341,6 +374,134 @@ def test_friend_request_flow(db):
         assert res.user_ids[0] == user2.id
 
 
+def test_SendFriendRequest_invisible_user_as_sender(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user(is_deleted=True)
+    with api_session(token2) as api:
+        with pytest.raises(grpc.RpcError) as e:
+            api.SendFriendRequest(
+                api_pb2.SendFriendRequestReq(
+                    user_id=user1.id,
+                )
+            )
+    assert e.value.code() == grpc.StatusCode.NOT_FOUND
+    assert e.value.details() == errors.USER_NOT_FOUND
+
+
+def test_SendFriendRequest_invisible_user_as_recipient(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user(is_deleted=True)
+    with api_session(token1) as api:
+        with pytest.raises(grpc.RpcError) as e:
+            api.SendFriendRequest(
+                api_pb2.SendFriendRequestReq(
+                    user_id=user2.id,
+                )
+            )
+    assert e.value.code() == grpc.StatusCode.NOT_FOUND
+    assert e.value.details() == errors.USER_NOT_FOUND
+
+
+"""""
+def test_ViewFriendRequest_invisible_user_as_sender(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user()
+
+    with api_session(token2) as api:
+        friend_request_id = api.SendFriendRequest(api_pb2.SendFriendRequestReq(user_id=user1.id)).friend_request_id
+
+    with session_scope() as session:
+        user2 = get_user_by_field(session, user2.username)
+        user2.is_banned = True
+        session.commit()
+        session.refresh(user2)
+        session.expunge(user2)
+
+    with api_session(token1) as api:
+        with pytest.raises(grpc.RpcError) as e:
+            api.GetFriendRequest(
+                api_pb2.GetFriendRequestReq(
+                    friend_request_id=friend_request_id,
+                )
+            )
+    assert e.value.code() == grpc.StatusCode.NOT_FOUND
+    assert e.value.details() == errors.FRIEND_REQUEST_NOT_FOUND
+
+
+def test_ViewFriendRequest_invisible_user_as_recipient(db):
+    # TODO
+    pass
+"""
+
+
+def test_ListFriendRequests_invisible_user_as_sender(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user()
+
+    # Check no active FR to start
+    with api_session(token1) as api:
+        res = api.ListFriendRequests(empty_pb2.Empty())
+        assert len(res.sent) == 0
+        assert len(res.received) == 0
+
+    # Send friend request to user 1
+    with api_session(token2) as api:
+        friend_request_id = api.SendFriendRequest(api_pb2.SendFriendRequestReq(user_id=user1.id)).friend_request_id
+
+    # Check 1 received FR
+    with api_session(token1) as api:
+        res = api.ListFriendRequests(empty_pb2.Empty())
+        assert len(res.sent) == 0
+        assert len(res.received) == 1
+
+    # Hide sender
+    with session_scope() as session:
+        user2 = get_user_by_field(session, user2.username)
+        user2.is_banned = True
+        session.commit()
+        session.refresh(user2)
+        session.expunge(user2)
+
+    # Check back to no FR
+    with api_session(token1) as api:
+        res = api.ListFriendRequests(empty_pb2.Empty())
+        assert len(res.sent) == 0
+        assert len(res.received) == 0
+
+
+def test_ListFriendRequests_invisible_user_as_recipient(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user()
+
+    # Check no active FR to start
+    with api_session(token1) as api:
+        res = api.ListFriendRequests(empty_pb2.Empty())
+        assert len(res.sent) == 0
+        assert len(res.received) == 0
+
+        # Send FR from user1
+        api.SendFriendRequest(api_pb2.SendFriendRequestReq(user_id=user2.id))
+
+        # Check one FR sent
+        res = api.ListFriendRequests(empty_pb2.Empty())
+        assert len(res.sent) == 1
+        assert len(res.received) == 0
+
+    # Hide recipient
+    with session_scope() as session:
+        user2 = get_user_by_field(session, user2.username)
+        user2.is_banned = True
+        session.commit()
+        session.refresh(user2)
+        session.expunge(user2)
+
+    # Check back to no FR
+    with api_session(token1) as api:
+        res = api.ListFriendRequests(empty_pb2.Empty())
+        assert len(res.sent) == 0
+        assert len(res.received) == 0
+
+
 def test_cant_friend_request_twice(db):
     user1, token1 = generate_user("user1")
     user2, token2 = generate_user("user2")
@@ -429,6 +590,28 @@ def test_ListFriends(db):
         assert user3.id in res.user_ids
 
 
+def test_ListFriends_with_invisible_users(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user(is_deleted=True)
+    user3, token3 = generate_user()
+
+    with api_session(token1) as api:
+        res = api.ListFriends(empty_pb2.Empty())
+        assert len(res.user_ids) == 0
+
+    make_friends(user1, user2)
+
+    with api_session(token1) as api:
+        res = api.ListFriends(empty_pb2.Empty())
+        assert len(res.user_ids) == 0
+
+    make_friends(user1, user3)
+
+    with api_session(token1) as api:
+        res = api.ListFriends(empty_pb2.Empty())
+        assert len(res.user_ids) == 1
+
+
 def test_mutual_friends(db):
     user1, token1 = generate_user("user1")
     user2, token2 = generate_user("user2")
@@ -505,6 +688,48 @@ def test_CancelFriendRequest(db):
 
         res = api.ListFriendRequests(empty_pb2.Empty())
         assert len(res.sent) == 0
+
+
+def test_RespondFriendRequest_invisible_user_as_sender(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user()
+
+    with api_session(token2) as api:
+        friend_request_id = api.SendFriendRequest(api_pb2.SendFriendRequestReq(user_id=user1.id)).friend_request_id
+
+    with session_scope() as session:
+        user2 = get_user_by_field(session, user2.username)
+        user2.is_banned = True
+        session.commit()
+        session.refresh(user2)
+        session.expunge(user2)
+
+    with api_session(token1) as api:
+        with pytest.raises(grpc.RpcError) as e:
+            api.RespondFriendRequest(api_pb2.RespondFriendRequestReq(friend_request_id=friend_request_id, accept=True))
+    assert e.value.code() == grpc.StatusCode.NOT_FOUND
+    assert e.value.details() == errors.FRIEND_REQUEST_NOT_FOUND
+
+
+def test_CancelFriendRequest_invisible_user_as_recipient():
+    user1, token1 = generate_user()
+    user2, token2 = generate_user()
+
+    with api_session(token1) as api:
+        friend_request_id = api.SendFriendRequest(api_pb2.SendFriendRequestReq(user_id=user2.id)).friend_request_id
+
+    with session_scope() as session:
+        user2 = get_user_by_field(session, user2.username)
+        user2.is_banned = True
+        session.commit()
+        session.refresh(user2)
+        session.expunge(user2)
+
+    with api_session(token1) as api:
+        with pytest.raises(grpc.RpcError) as e:
+            api.CancelFriendRequest(api_pb2.CancelFriendRequestReq(friend_request_id=friend_request_id))
+    assert e.value.code() == grpc.StatusCode.NOT_FOUND
+    assert e.value.details() == errors.FRIEND_REQUEST_NOT_FOUND
 
 
 def test_reject_friend_request(db):
@@ -623,6 +848,36 @@ def test_reporting(db):
     with api_session(token1) as api:
         with pytest.raises(grpc.RpcError) as e:
             api.Report(report_req)
+    assert e.value.code() == grpc.StatusCode.NOT_FOUND
+    assert e.value.details() == errors.USER_NOT_FOUND
+
+
+def test_invisible_user_writes_reference(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user(accepted_tos=0)
+
+    with api_session(token2) as api:
+        with pytest.raises(grpc.RpcError) as e:
+            api.WriteReference(
+                api_pb2.WriteReferenceReq(
+                    to_user_id=user1.id, reference_type=api_pb2.ReferenceType.FRIEND, text="smells funny"
+                )
+            )
+    assert e.value.code() == grpc.StatusCode.NOT_FOUND
+    assert e.value.details() == errors.USER_NOT_FOUND
+
+
+def test_write_reference_for_invisible_user(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user(accepted_tos=0)
+
+    with api_session(token1) as api:
+        with pytest.raises(grpc.RpcError) as e:
+            api.WriteReference(
+                api_pb2.WriteReferenceReq(
+                    to_user_id=user2.id, reference_type=api_pb2.ReferenceType.FRIEND, text="smells funny"
+                )
+            )
     assert e.value.code() == grpc.StatusCode.NOT_FOUND
     assert e.value.details() == errors.USER_NOT_FOUND
