+++ app/backend/src/tests/test_search.py
@@ -1,5 +1,6 @@
 import grpc
 import pytest
+from google.protobuf import wrappers_pb2
 
 from couchers import errors
 from couchers.db import session_scope
@@ -44,3 +45,26 @@ def test_UserSearch(testing_communities):
     with search_session(token) as api:
         res = api.UserSearch(search_pb2.UserSearchReq())
         assert len(res.results) > 0
+
+
+def test_Search_function_invisible_users(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user(is_deleted=True)
+
+    with search_session(token1) as api:
+        res = api.Search(
+            search_pb2.SearchReq(
+                query="test_user_",
+                include_users=True,
+            )
+        )
+        assert len(res.results) == 1
+
+
+def test_UserSearch_function_invisible_users(db):
+    user1, token1 = generate_user()
+    user2, token2 = generate_user(is_deleted=True)
+
+    with search_session(token1) as api:
+        res = api.UserSearch(search_pb2.UserSearchReq(query=wrappers_pb2.StringValue(value="test_user_")))
+        assert len(res.results) == 1
